requirements.txt
accelerate==0.33.0
transformers==4.44.0
label_studio_tools
prometheus-client
label_studio_tools
aixblock-ml
aixblock-sdk
aixblock-tools
tensorboardX
gradio #==3.50
redis
rq
flask>=2.2
flask-cors
Werkzeug>=2.0
gunicorn>=20.0
colorama
lxml
appdirs
gitpython
centrifuge-python
opencv-python>=4.1.1
numpy
grafana-api
diffusers
datasets
trl
peft
scipy
bitsandbytes>=0.41.3
langchain
langchain_community
nvidia-ml-py
wandb
langchain_huggingface
flask_swagger_ui
pyyaml
gradio_log
tf-keras
loguru
about-time
rich
sparkdl
validators
triton==3.1.0

opencv-python==4.8.1.78
einops==0.7.0
pytorch-lightning==1.9.0
# bitsandbytes==0.44.0
prodigyopt==1.0
lion-pytorch==0.0.6
schedulefree==1.4

safetensors==0.4.4

altair==4.2.2
easygui==0.98.3
toml==0.10.2
voluptuous==0.13.1
huggingface-hub==0.25.2
# for Image utils
imagesize==1.4.1

sentencepiece==0.2.0
mcp
torch==2.6.0 --index-url https://download.pytorch.org/whl/cu126
torchvision==0.21.0 --index-url https://download.pytorch.org/whl/cu126
torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu126